Summary of issue:
The current ingress.yaml file is using a deprecated way to declare the ingress class as 'nginx.' This caused a validation error for the ingress-ngnix controller pod. The controller could not validate which ingress to use.

We updated the ingress.yaml to declare the ingress class using the following addition under the spec section. This is the updated way to declare an ingress class.

This resolved the ingress routing issue.